Add Constant mean function, revive mean_functions.py
