+++ gptorch/__init__.py
@@ -1,5 +1,12 @@
+# File: __init__.py
+# Author: Yinhao Zhu (yzhu10@nd.edu)
 
-from __future__ import absolute_import
-from . import (functions, model, kernels, ekernels, likelihoods, mean_functions,
-               densities, util)
+from . import model
+from . import functions
+from . import mean_functions
+from . import kernels
+from . import ekernels
+from . import likelihoods
+from . import densities
+from . import util
 from . import models
