+++ gptorch/models/__init__.py
@@ -1,4 +1,6 @@
-from __future__ import absolute_import
+# File: __init__.py
+# Author: Yinhao Zhu (yzhu10@nd.edu)
+
+from . import base
 from .gpr import GPR
 from .sparse_gpr import VFE, SVGP
-
